Email validation was implemented so that when the user has already registered, he/she only logs in and no more users are created in the database.
